final - done all i could
what's working is the creation of a .configure file as well as the
multithreaded wtf server.  the client executable is able to connect to
the server and the server handles that client's connection in its own
thread.  the check configure method works and pulls out the ip as well
as the port from the configure file and is used when connecting the
client to the server.  All other functionalities are left out. the
amount of work required to create them and implement them into this
project is rather obscene so quite honestly fuck that. gg.